requirements.txt
numpy==2.4.4
pandas==3.0.2
patsy==1.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
scipy==1.17.1
statsmodels==0.14.6
streamlit>=1.30.0
plotly>=5.18.0
